requirements.txt
pyserial>=3.5
tkinter
Pillow>=8.0.0
matplotlib>=3.0.0
numpy>=1.19.0
pandas>=1.1.0
gitpython>=3.1.0
